requirements.txt
# Appium and Selenium
Appium-Python-Client>=3.1.0
selenium>=4.15.0

# Testing Framework
pytest>=7.4.0
pytest-html>=3.2.0
pytest-xdist>=3.3.0

# Utilities
python-dotenv>=1.0.0
